(+) Check valid id

+++ backend/src/api/services/postService.js
@@ -1,6 +1,7 @@
 import { Post, User } from '../models';
 import httpStatus from 'http-status-codes';
-import HTTPError from '../helpers/class/httpErrors';
+import HTTPError from '../helpers/classes/httpErrors';
+import mongoose from 'mongoose';
 
 export default {
     getPosts: async () => {
@@ -14,6 +15,12 @@ export default {
     },
 
     addPost: async (data) => {
+        if (!mongoose.isValidObjectId(data.owner)) {
+            throw new HTTPError(httpStatus.UNPROCESSABLE_ENTITY, {
+                title: 'Unprocessable entity',
+                detail: 'Invalid user id'
+            });
+        }
         let user = await User.findById(data.owner);
         if (user === null) {
             throw new HTTPError(httpStatus.NOT_FOUND, {
@@ -27,6 +34,12 @@ export default {
     },
 
     getPost: async (postId) => {
+        if (!mongoose.isValidObjectId(postId)) {
+            throw new HTTPError(httpStatus.UNPROCESSABLE_ENTITY, {
+                title: 'Unprocessable entity',
+                detail: 'Invalid post id'
+            });
+        }
         let post = await Post.findById(postId, (err) => {
             if (err) throw err;
         })
@@ -45,6 +58,12 @@ export default {
     },
 
     updatePost: async (postId, data) => {
+        if (!mongoose.isValidObjectId(postId)) {
+            throw new HTTPError(httpStatus.UNPROCESSABLE_ENTITY, {
+                title: 'Unprocessable entity',
+                detail: 'Invalid post id'
+            });
+        }
         let post = await Post.findByIdAndUpdate(postId, data, { new: true });
         if (post === null) {
             throw new HTTPError(httpStatus.NOT_FOUND, {
@@ -56,6 +75,12 @@ export default {
     },
 
     deletePost: async (postId) => {
+        if (!mongoose.isValidObjectId(postId)) {
+            throw new HTTPError(httpStatus.UNPROCESSABLE_ENTITY, {
+                title: 'Unprocessable entity',
+                detail: 'Invalid post id'
+            });
+        }
         let post = await Post.findByIdAndDelete(postId);
         if (post === null) {
             throw new HTTPError(httpStatus.NOT_FOUND, {
